fix NULL dereference on wxWidgets 3.2
According to the document of wxWidgets, the initializer of wxLocale with
no parameters do not do any initialization at all, thus do not register
its wxTranslations.

Explicitly call Init() with its default parameters to init it with the
default setting, otherwise on Linux with wxWidgets 3.2, the
wxTranslations singleton get is NULL and lead to NULL deference.

The code here is from PrusaSlicer 2.5 [1].

[1] https://github.com/prusa3d/PrusaSlicer/commit/e21921f2ebd2b94c57459bd280f1b1bd8eec1b9c

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>